requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
rich>=13.7.0
click>=8.1.0
